Cleanup of Orleans logging layer. (#1844)
TraceLogger has been broken into three classes and deprecated. It is still maintained mainly to support programmatic log configuration until customers can migrate to the new pattern.
 - LoggerImpl replaces TraceLoggers logger implementation role. It is internal and all public api surfaces now use Logger.
 - LogManager replaces TraceLoggers support of programmatic logging settings.
 - LogFormatter replaces TraceLoggers log formatting utility functions.

+++ src/Orleans/Async/TaskExtensions.cs
@@ -171,7 +171,7 @@ namespace Orleans
             catch (Exception exc)
             {
                 var ignored = task.Exception; // Observe exception
-                logger.Error((int)errorCode, message, exc);
+                logger.Error(errorCode, message, exc);
                 throw;
             }
         }
