+++ src/Orleans/Streams/SimpleMessageStream/SimpleMessageStreamProducerExtension.cs
@@ -227,7 +227,7 @@ namespace Orleans.Providers.Streams.SimpleMessageStream
                     if (consumers.TryRemove(subscriptionId, out discard))
                     {
                         streamPubSub.UnregisterConsumer(subscriptionId, streamId, streamId.ProviderName).Ignore();
-                        logger.Warn((int)ErrorCode.Stream_ConsumerIsDead,
+                        logger.Warn(ErrorCode.Stream_ConsumerIsDead,
                             "Consumer {0} on stream {1} is no longer active - permanently removing Consumer.", remoteConsumer, streamId);
                     }
                 }
